refactor: WinningNumber 리패키징
