+++ src/main/java/lotto/dto/WinningNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.dto;
 
 import java.util.Arrays;
 import java.util.List;
